Major updates: removing support for Blaze library; change EigenNLEigenSolver to JacobiDavidsonNLEigensolver; Added inverseFreeKrylovNLEigenSolver.

+++ NLEigenSolver/src/Instrumentor.h
@@ -14,7 +14,7 @@ struct ProfileResult
 {
 	std::string Name;
 	long long Start, End;
-	uint32_t ThreadID;
+	size_t ThreadID;
 };
 
 struct InstrumentationSession
@@ -117,7 +117,7 @@ public:
 
 		float duration = (end - start) * 0.001f;
 		//std::cout << m_Name << ": " << duration << "ms" << std::endl;
-		uint32_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
+		size_t threadID = std::hash<std::thread::id>{}(std::this_thread::get_id());
 		Instrumentor::Get().WriteProfile({ m_Name, start, end , threadID });
 
 		m_Stopped = true;
